pyproject.toml
[project]
name = "pediatric"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "numpy==2.2.2",
    "ollama==0.4.7",
    "openai==1.60.2",
    "python-dotenv==1.0.1",
    "qdrant-client==1.13.2",
]

[tool.ruff]
line-length = 80
